Changes: Create post button to Manage post feature
